requirements.txt
fastapi
python-dotenv
httpx
requests